+++ src/main/java/com/fi/springboot/onlineparkingsystem/service/RequestService.java
@@ -11,8 +11,8 @@ import com.fi.springboot.onlineparkingsystem.entity.Request;
 public interface RequestService
 {
 	List<Request> getAllRequest();
-	Request getRequestById(int id);
+	Request getRequestById(long id);
 	public Request registerRequest(Request request);
-	public ResponseEntity<Request> updateRequest(int id,Request requestDetails);
-	public String deleteRequest(int id);
+	public ResponseEntity<Request> updateRequest(long id,Request requestDetails);
+	public String deleteRequest(long id);
 }
